Node file to store data for linkedlist
Why this change is needed:
This is the initial push to master containing a node that will
be used to store data for the linked list that will be made.

What this change accomplishes:
Without the node the the linked list doesn't have a container for data.

Issue Number:

Related Issues: